% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datacomb.R
\name{renderDatacomb}
\alias{renderDatacomb}
\title{Widget render function for use in Shiny}
\usage{
renderDatacomb(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

