% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datacomb.R
\name{Datacomb}
\alias{Datacomb}
\title{Datacomb}
\usage{
Datacomb(dataFrame, columns = c(), rowLabel = NULL, width = "100\%",
  height = "100\%")
}
\description{
An interface for viewing and combing through data frames.
}
\examples{
library(datacomb)

df <- data.frame(model=rownames(mtcars),mtcars)
Datacomb(
  df,
  columns = colnames(df)[-1],
  rowLabel = "model"
)
}

