% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datacomb.R
\name{Datacomb}
\alias{Datacomb}
\title{Datacomb}
\usage{
Datacomb(data = NULL, columns = colnames(data), rowLabel = NULL,
  width = NULL, height = NULL)
}
\arguments{
\item{data}{a data object of a class that can be converted into a
\code{data.frame}}

\item{columns}{names of the columns in \code{data}
to use as dimensions}

\item{rowLabel}{name of the column in \code{data}
that provides the row label}

\item{width,height}{valid \code{CSS} size unit for the
         width and height of the datacomb.}
}
\description{
An interactive interface for viewing and combing through data frames.
}
\examples{
library(datacomb)

Datacomb(mtcars)

data(diamonds, package = "ggplot2")
Datacomb(diamonds)
}

